Looks like the login component throws an error when it is missing the errors property
This should actually be defaulted to something. I also noticed that we are logging into the console when there are prop-type errors. This is most likely the Winston configuration. I have not expose the configuration as an API but this seems like a good use-case for leverage external logging infrastructure (e.g. New Relic).